Move form data cleaning to JSONSchemaForm component
